% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERhjelpefunksjoner.R
\name{abonnementNGER}
\alias{abonnementNGER}
\title{Kjøre samlerapporter for abonnement i NGER}
\usage{
abonnementNGER(
  rnwFil,
  brukernavn = "ngerBrukernavn",
  reshID = 0,
  datoFra = Sys.Date() - 180,
  datoTil = Sys.Date()
)
}
\arguments{
\item{rnwFil}{NoWeb-fil med filending: eksempel.rnw}

\item{brukernavn}{innlogget bruker}

\item{reshID}{brukerens reshID}

\item{datoFra}{startdato}

\item{datoTil}{sluttdato}
}
\value{
gir filsti til pdf-samledokument
}
\description{
Kjøre samlerapporter for abonnement i NGER
}
