% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERhjelpefunksjoner.R
\name{dataTilResPort}
\alias{dataTilResPort}
\title{Generere data til Resultatportalen}
\usage{
dataTilResPort(
  RegData = RegData,
  valgtKI,
  datoFra = "2016-01-01",
  figurtype = "andelGrVar",
  aar = 2016:2019,
  OpMetode = 0,
  filUt = "dummy"
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtKI}{- KomplIntra, KomplPostop, Opf0AlvorlighetsGrad,
HysKonvertert, LapKonvertert, Tss2Generelt, Tss2Sumskaar}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{OpMetode}{1: Laparoskopi
2: Hysteroskopi
3: Begge
4: LCD01 eller LCD04 (total laparoskopisk hysterektomi)
5: LCC11 (laparoskopisk subtotal hysterektomi)
6: LCD11 (laparoskopisk assistert vaginal hysterektomi)}

\item{filUt}{tilnavn for utdatatabell (fjern?)}
}
\value{
Datafil til Resultatportalen
}
\description{
Generere data til Resultatportalen
}
