% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NGERFigAndeler.R
\name{FigAndeler}
\alias{FigAndeler}
\title{Lag søylediagram eller stabelplott som viser andeler av ulike variabler}
\usage{
FigAndeler(RegData, valgtVar, datoFra = "2000-04-01",
  datoTil = "2050-12-31", minald = 0, maxald = 130, outfile = "",
  reshID, enhetsUtvalg = 1, MCEType = 99, hentData = F, preprosess = T)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel skal plottes}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Default)
2: Egen enhet}

\item{MCEType}{1: Laparoskopi
                2: Hysteroskopi
                3: Begge
                99: Alle}

\item{hentData}{Gjør spørring mot database
FALSE: Nei, RegData gis som input til funksjonen (Default)
TRUE: Ja}

\item{preprosess}{Preprosesser data
FALSE: Nei (Default)
TRUE: Ja}
}
\value{
En figur med søylediagram eller et stabelplot av ønsket variabel
}
\description{
Denne funksjonen lager et søylediagram eller stabelplot som viser andeler av valgt variabel
filtrert på de utvalg som er gjort.
}

