% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERFigAndelerGrVar.R
\name{FigAndelerGrVar}
\alias{FigAndelerGrVar}
\title{Søylediagram med andeler for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
FigAndelerGrVar(RegData = 0, valgtVar, datoFra = "2013-01-01",
  datoTil = "3000-12-31", minald = 0, maxald = 130, MCEType = 99,
  reshID, outfile = "", enhetsUtvalg = 1, preprosess = TRUE,
  hentData = 0)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Alder: Pasienter over 70 år
FollowupSeriousness: Andel av postoperative komplikasjoner som var alvorlige (3 og 4)
KomplIntra: Komplikasjoner ved operasjon. (kombinerer variablene HypComplications og LapComplications)
KomplPostop: Andel postoperative komplikasjoner
OpAntibioticProphylaxis: Andel som får antibiotika
OpBMI: Pasienter med fedme (BMI>30)
Reop: Andel reoperasjon som følge av komplikasjon
StatusFollowup: Pasienter som har fått postoperativ oppfølging}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{MCEType}{1: Laparoskopi
                2: Hysteroskopi
                3: Begge
                99: Alle

Detajer: Her bør man liste opp hvilke variable funksjonen benytter.}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Standard)
2: Egen enhet}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{valgtVar:}{Velg hvilken variabel du ønsker å se resultat for
Alder: Andel pasienter over 70 år.
ComplReop: Reoperasjon som følge av komplikasjon
Education: Pasienter med høyere utdanning
FollowupSeriousness: Alvorlige komplikasjoner
KomplIntra: Komplikasjoner under operasjon (intraoperativt)
KomplPostop: Postoperative komplikasjoner
OpAntibioticProphylaxis: Fått antibiotikaprofylakse
OpASA: ASA-grad > II
OpBMI: Pasienter med fedme (BMI>30)
StatusFollowup: Fått postoperativ pensum}
}
\description{
Funksjon som genererer en figur med andeler av en variabel for en grupperingsvariabelen sykehus.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus
}
\details{
Detajer: Variable funksjonen benytter: Alder (beregnes), ComplExist, ComplReop, Education, FollowupSeriousness
 HypComplications, LapComplications, MCEType, OpAntibioticProphylaxis, OpASA, OpBMI, StatusFollowup.
 Det benyttes også andre variable til utvalg osv.
}

