% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERFigAndelTid.R
\name{FigAndelTid}
\alias{FigAndelTid}
\title{Tidstrend (år) av andel for en gitt variabel.}
\usage{
FigAndelTid(RegData = 0, valgtVar, datoFra = "2013-01-01",
  datoTil = "3000-12-31", minald = 0, maxald = 130, MCEType = 99,
  Hastegrad = "", AlvorlighetKompl = "", reshID, outfile = "",
  enhetsUtvalg = 1, preprosess = TRUE, hentData = 0)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Alder: Pasienter over 70 år
ComplReop: Andel reoperasjon som følge av komplikasjon
FollowupSeriousness: Andel av postoperative komplikasjoner som var alvorlige (3 og 4)
KomplIntra: Komplikasjoner ved operasjon. (kombinerer variablene HypComplications og LapComplications)
KomplPostop: Andel postoperative komplikasjoner
OpAntibioticProphylaxis: Andel som får antibiotika
OpBMI: Pasienter med fedme (BMI>30)
StatusFollowup: Pasienter som har fått postoperativ oppfølging}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{MCEType}{1: Laparoskopi
2: Hysteroskopi
3: Begge
99: Alle}

\item{Hastegrad}{Hastegrad av operasjon.
1: Elektiv
2: Akutt
3: Ø-hjelp}

\item{AlvorlighetKompl}{Alvorlighetsgrad for postoperative komplikasjoner (Flervalg)
                         Angis som en vektor av tall som tekst, f.eks. c('1','2')
                         1: Lite alvorlig
                         2: Middels alvorlig
                         3: Alvorlig
                         4: Dødelig

Detajer: Her bør man liste opp hvilke variable funksjonen benytter.}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Standard)
2: Egen enhet}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}
}
\description{
Funksjon som genererer en figur med andeler av en variabel for hvert år.
}
\details{
Detaljer:
}

