% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERhjelpefunksjoner.R
\name{henteSamlerapporter}
\alias{henteSamlerapporter}
\title{Generere samlerapporter i app}
\usage{
henteSamlerapporter(filnavn, rmdFil, reshID = 0)
}
\arguments{
\item{filnavn}{benyttes i downloadhandler}

\item{rmdFil}{Fila som skal kompileres. Eks. 'Eksempel.Rmd'}

\item{reshID}{reshID}
}
\description{
Generere samlerapporter i app
}
