% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERhjelpefunksjoner.R
\name{hjelpeFunksjoner}
\alias{hjelpeFunksjoner}
\alias{SorterOgNavngiTidsEnhet}
\alias{lageTulleData}
\title{Hjelpefunksjoner. Group of functions page title}
\usage{
SorterOgNavngiTidsEnhet(RegData, tidsenhet = "Aar", tab = 0)

lageTulleData(RegData, varBort = "", antSh = 26, antObs = 20000)
}
\arguments{
\item{RegData}{data}

\item{PasientID}{Variabelen som angir pasientidentifikasjon}
}
\value{
Div hjelpefunksjoner
}
\description{
Fil med div hjelpefunksjoner.Group of functions Description section
}
\details{
Detaljer. kommer senereGroup of functions Details paragraph.
}
\section{Finne reinnleggelser After function section}{

Despite its location, this actually comes after the function section.
Fil som inneholder hjelpefunksjoner.
Lage tulledata for åpen publisering
SorterOgNavngiTidsEnhet Legger til tidsenhetene Aar, Halvaar, Mnd og Kvartal
}

\section{Tilrettelegge tidsenhetvariabel}{

Probably better if all sections come first, uless have one section per function. Makes it easier to
see the information flow.
}

\section{Lage tulledata (simulerte data)}{
NA
}

