% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERtabeller.R
\name{instrumentbruk}
\alias{instrumentbruk}
\title{instrumentbruk, Laparoskopi}
\usage{
instrumentbruk(RegData, datoFra = "2019-01-01", datoTil = Sys.Date())
}
\description{
instrumentbruk, Laparoskopi
}
