% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERhjelpefunksjoner.R
\name{kjor_NGERapp}
\alias{kjor_NGERapp}
\title{Kjør Shiny Application}
\usage{
kjor_NGERapp(browser = FALSE, logAsJson = FALSE)
}
\arguments{
\item{browser}{App åpner i browser}

\item{logAsJson}{Logg i json-format}
}
\value{
Et objekt som representerer den NGERapp'en
}
\description{
Kjør Shiny Application
}
