% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERAntallRegPrAvd.R
\name{NGERAntallRegPrAvd}
\alias{NGERAntallRegPrAvd}
\title{Lag tabell med antall registreringer per avdeling}
\usage{
NGERAntallRegPrAvd(RegData = RegData, datoFra = "2000-01-01",
  datoTil = "2050-12-31")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}
}
\value{
Tabell En tabell over antall registreringer per år per avdeling
}
\description{
Denne funksjonen lager en tabell over antall registreringer per år per avdeling
}

