% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERFigAntReg.R
\name{NGERFigAntReg}
\alias{NGERFigAntReg}
\title{Søylediagram som viser antall registreringer per måned}
\usage{
NGERFigAntReg(
  RegData,
  datoTil = Sys.Date(),
  reshID = 0,
  minald = 0,
  maxald = 130,
  erMann = "",
  preprosess = 0,
  hentData = 0,
  outfile = ""
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{preprosess}{Preprosesser data
0: Nei (Standard)
1: Ja}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}
}
\value{
Søylediagram som viser antall registreringer per måned for valgte 12 måneder
}
\description{
Søylediagram som viser antall registreringer per måned, siste 12 hele måneder fra valgt sluttdato,
 og sammenligner med fjoråret.
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}
