% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERHentData.R
\name{NGERHentRegData}
\alias{NGERHentRegData}
\title{Provide global dataframe for NGER}
\usage{
NGERHentRegData(datoFra = "2014-01-01", datoTil = "2099-01-01")
}
\arguments{
\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}
}
\value{
RegData data frame
}
\description{
Provides NGER data from staging
}
