% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERqueryVI.R
\name{NGERHentRegDataVI}
\alias{NGERHentRegDataVI}
\title{Provide dataframe for tab vi NGER}
\usage{
NGERHentRegDataVI(reportYear = 2099)
}
\value{
RegData data frame
}
\description{
Provides NGER data for tab vi from staging
}
