% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERqueryXXI.R
\name{NGERHentRegDataXXI}
\alias{NGERHentRegDataXXI}
\title{Provide dataframe for tab xxi NGER}
\usage{
NGERHentRegDataXXI(datoFra = "2014-01-01", datoTil = "2099-01-01")
}
\arguments{
\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}
}
\value{
RegData data frame
}
\description{
Provides NGER data for tab xxi from staging
}
