% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NGEROpInd.R
\name{NGEROpInd}
\alias{NGEROpInd}
\title{Genererer figur over de 20 vanligste operasjonsindikatorene.}
\usage{
NGEROpInd(RegData = 0, datoFra = "2000-04-01", datoTil = "2050-12-31",
  minald = 0, maxald = 130, outfile = "", reshID, enhetsUtvalg = 1,
  MCEType = 99, preprosess = TRUE, hentData = TRUE)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Standard)
2: Egen enhet}

\item{MCEType}{1: Laparoskopi
                2: Hysteroskopi
                3: Begge
                99: Alle}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}
}
\value{
Tabell med de 20 vanligste operasjonsindikatorene
}
\description{
Denne funksjonen lager en figur over de 20 vanligste operasjonsindikatorene.
I tellingen slås variablene OpInd1, OpInd2 og OpInd3 sammen.
}

