% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAND36dimensjoner.R
\name{NGERRAND36sjekk}
\alias{NGERRAND36sjekk}
\title{Beregning og test av tidligere beregnede dimensjoner i RAND36}
\usage{
NGERRAND36sjekk(RegData)
}
\value{
RAND36
}
\description{
Beregning og test av tidligere beregnede dimensjoner i RAND36
}
