% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERregDelay.R
\name{NGERregDealy}
\alias{NGERregDealy}
\title{Provide registration delay for NGER}
\usage{
NGERregDealy(years)
}
\arguments{
\item{years}{integer vector with years for results and grouping}
}
\value{
data frame with registry name and values for each year
}
\description{
Provide registration delay in median number of days grouped by years
}

