% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERqueryRegDelay.R
\name{NGERRegDelayData}
\alias{NGERRegDelayData}
\title{Provide registration delay data from NGER}
\format{Return a data frame with two variables:
\describe{
\item{year}{the year for the event (from db field \emph{HovedDato})}
\item{daysDiff}{the difference in days between event (from db field
\emph{HovedDato}) and the last time the registration form was saved
(from db field \emph{SistLagretDato})}
}}
\usage{
NGERRegDelayData()
}
\value{
RegData data frame
}
\description{
Provides delay between event (main date) and registration as a data frame
}
\details{
For the query these conditions apply:
\describe{
\item{SkjemaStatus=1}{use Only finished registrations}
\item{SkjemaNavn="Operasjon"}{use only registry form 'Operasjon'}
}
}

