% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERgmlAre.R
\name{NGERtabVI}
\alias{NGERtabVI}
\title{Generate tab VI}
\usage{
NGERtabVI(RegData)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}
}
\value{
list $tabVI data frame of table data
}
\description{
Yes, generate tab VI
}
