% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERVarTilrettelegg.R
\name{NGERVarTilrettelegg}
\alias{NGERVarTilrettelegg}
\title{Funksjon for å tilrettelegge variable for beregning.}
\usage{
NGERVarTilrettelegg(
  RegData,
  valgtVar,
  OpMetode = 0,
  ind = 0,
  figurtype = "andeler"
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres}

\item{OpMetode}{1: Laparoskopi
 2: Hysteroskopi
 3: Begge # 21.okt. 2022: Fjernes som eget valg. Blir med i både laparoskopi og hysteroskopi
 4: LCD01 eller LCD04 (total laparoskopisk hysterektomi)
 5: LCC11 (laparoskopisk subtotal hysterektomi)
 6: LCD11 (laparoskopisk assistert vaginal hysterektomi)
 7: Lap m/robot
 8: Kolpopeksiene
 9: Hysterectomier (alle)
10: Lap u/robot
11: Laparoskopi med vaginal tilgang}

\item{ind}{indekser fra enhetsUtvalg. Benyttes normalt ikke her, men trengs for
valgtVar Diagnoser, DiagnoseGr, Prosedyrer og ProsedyreGr}

\item{figurtype}{Hvilken figurtype det skal tilrettelegges variable for:
'andeler', 'andelGrVar', 'andelTid', 'gjsnGrVar', 'gjsnTid'}
}
\value{
Definisjon av valgt variabel.
}
\description{
Denne funksjonen gjør utvalg og tilrettelegger variable (gitt ved valgtVar) til videre bruk.
Videre bruk kan eksempelvis være beregning av AggVerdier eller gjennomsnitt.
Funksjonen gjør også filtreringer som å fjerne ugyldige verdier for den valgte variabelen, samt ta høyde for avhengigheter med
andre variable. Det er også her man angir aksetekster og titler for den valgte variabelen.
Her kan mye hentes til analysebok
}
