% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERFigAndelerGrVar.R
\name{PlotAndelerGrVar}
\alias{PlotAndelerGrVar}
\title{Andel per enhet (eller annen grupperingsvariabel)}
\usage{
PlotAndelerGrVar(
  RegData,
  AggVerdier,
  Ngr,
  N,
  GrNavnSort,
  andeltxt,
  grVar = "ShNavn",
  hovedgrTxt = "",
  KvalIndGrenser = NA,
  tittel = "tittel",
  utvalgTxt = "",
  Ngrense = 10,
  titleSize = 20,
  subtitleSize = 15,
  legendSize = 12,
  axisTextSize = 10,
  sortAvtagende = TRUE,
  nTicks = 5,
  fargepalett = "BlaaOff",
  outfile = ""
)
}
\arguments{
\item{RegData}{dataramme}

\item{AggVerdier}{aggregerte tall}

\item{grVar}{grvar}

\item{hovedgrTxt}{hovedtxt}

\item{KvalIndGrenser}{KIgrenser}

\item{tittel}{tittel}

\item{Ngrense}{minste gruppeantall som vises}

\item{titleSize}{titleSize}

\item{subtitleSize}{subtitleSize}

\item{legendSize}{legendSize}

\item{axisTextSize}{axisTextSize}

\item{nTicks}{nTicks}

\item{fargepalett}{fargepalett}

\item{outfile}{filtype ut}
}
\value{
Figur
}
\description{
Andel per enhet (eller annen grupperingsvariabel)
}
