% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appNGER.R
\name{server_nger}
\alias{server_nger}
\title{Server-del til appen}
\usage{
server_nger(input, output, session)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\value{
Server-delen til NGER-appen
}
\description{
Server-del til appen
}
