% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlagFigAntStabel.R
\name{SlagFigAntReg}
\alias{SlagFigAntReg}
\title{Søylediagram som viser antall registreringer per måned}
\usage{
SlagFigAntReg(
  RegData,
  datoTil = "2050-12-31",
  minald = 0,
  maxald = 130,
  erMann = "",
  diagnose = "",
  innl4t = "",
  NIHSSinn = "",
  outfile = "",
  reshID,
  enhetsUtvalg = 2,
  preprosess = 1,
  hentData = 0
)
}
\arguments{
\item{enhetsUtvalg}{Gjør gruppeutvalg for
 0: Hele landet
 2: Egen enhet
7: Egen region}
}
\value{
Søylediagram som viser antall registreringer per måned for valgte 12 måneder
}
\description{
Søylediagram som viser antall registreringer per måned, siste 12 hele måneder fra valgt sluttdato, 
 av hovedskjema, samt oppfølgingsskjema+døde
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}
