% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERhjelpefunksjoner.R
\name{SorterOgNavngiTidsEnhet}
\alias{SorterOgNavngiTidsEnhet}
\title{Tilrettelegge tidsenhetvariabel:
Probably better if all sections come first, uless have one section per function. Makes it easier to
see the information flow.}
\usage{
SorterOgNavngiTidsEnhet(RegData, tidsenhet = "Aar", tab = 0)
}
\arguments{
\item{RegData}{dataramme}

\item{tidsenhet}{tidsenhet: 'Aar', 'Mnd', 'Kvartal', 'Halvaar'}

\item{tab}{Hmmm}
}
\description{
Tilrettelegge tidsenhetvariabel:
Probably better if all sections come first, uless have one section per function. Makes it easier to
see the information flow.
}
