% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERtabeller.R
\name{tabAntOpphSh5Aar}
\alias{tabAntOpphSh5Aar}
\title{Antall opphold siste 5 år}
\usage{
tabAntOpphSh5Aar(RegData, datoTil = Sys.Date(), OpMetode = 99, velgDiag = 0)
}
\description{
Antall opphold siste 5 år
}
