% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERtabeller.R
\name{tabAntOpphShMnd}
\alias{tabAntOpphShMnd}
\title{Fil som inneholder funksjoner for å lage tabeller, i første rekke tellinger av personer
RegData må inneholde InnDato og Aar.
-tabAntOpphSh12mnd: Antall opphold per måned og enhet siste 12 måneder fram til datoTil.
-tabAntOpphSh5Aar:Antall opphold per år og enhet siste 5 år (inkl. inneværende år) fram til datoTil.
Antall opphold siste X (antMnd) mnd}
\usage{
tabAntOpphShMnd(
  RegData,
  datoTil = Sys.Date(),
  antMnd = 6,
  reshID = 0,
  OpMetode = 99,
  velgDiag = 0
)
}
\arguments{
\item{RegData}{data}

\item{datoTil}{sluttdato. Brukes i tabellene AntOpph per 12 mnd og Belegg}

\item{personIDvar}{Variabelen som angir pasientidentifikasjon}
}
\value{
Div tabeller
}
\description{
Fil som inneholder funksjoner for å lage tabeller, i første rekke tellinger av personer
RegData må inneholde InnDato og Aar.
-tabAntOpphSh12mnd: Antall opphold per måned og enhet siste 12 måneder fram til datoTil.
-tabAntOpphSh5Aar:Antall opphold per år og enhet siste 5 år (inkl. inneværende år) fram til datoTil.
Antall opphold siste X (antMnd) mnd
}
