% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERtabeller.R
\name{tabAntSkjemaGml}
\alias{tabAntSkjemaGml}
\title{Hvor mange skjema av hver type}
\usage{
tabAntSkjemaGml(
  skjemaoversikt,
  datoFra = "2019-01-01",
  datoTil = Sys.Date(),
  skjemastatus = 1
)
}
\description{
Hvor mange skjema av hver type
}
