% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERtabeller.R
\name{tabKomplLap}
\alias{tabKomplLap}
\title{komplikasjoner, Laparaskopi}
\usage{
tabKomplLap(RegData, reshID = 0, datoFra = "2019-01-01", datoTil = Sys.Date())
}
\description{
komplikasjoner, Laparaskopi
}
