% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERtabeller.R
\name{tabKonvertertLap}
\alias{tabKonvertertLap}
\title{Konvertert laparoskopi til laparotomi}
\usage{
tabKonvertertLap(
  RegData,
  reshID = 0,
  datoFra = "2016-01-01",
  datoTil = Sys.Date()
)
}
\arguments{
\item{RegData}{dataramme}

\item{reshID}{reshID}

\item{datoFra}{startdato}

\item{datoTil}{sluttdato}
}
\description{
Konvertert laparoskopi til laparotomi
}
