% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERtabeller.R
\name{tabNGERpasientegenskaper}
\alias{tabNGERpasientegenskaper}
\title{Generere tabell med nøkkeltall}
\usage{
tabNGERpasientegenskaper(
  RegData,
  datoFra = "2019-01-01",
  datoTil = Sys.Date(),
  tidsenhet = "Kvartal"
)
}
\description{
Generere tabell med nøkkeltall
}
