% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERtabeller.R
\name{tabNokkelHys}
\alias{tabNokkelHys}
\title{Nøkkeltallstabell for hysteroskopi}
\usage{
tabNokkelHys(RegData = RegData)
}
\arguments{
\item{RegData}{NGER-data, dataramme}
}
\value{
Leverer formattert tabell for nøkkeltall
}
\description{
Nøkkeltallstabell for hysteroskopi
}
