% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERtabeller.R
\name{tabNokkelLap}
\alias{tabNokkelLap}
\title{Nøkkeltallstabell for laparoskopi}
\usage{
tabNokkelLap(
  RegData = RegData,
  datoFra = Sys.Date() - 365,
  datoTil = Sys.Date(),
  reshID = 0,
  velgAvd = 0,
  enhetsUtvalg = 0
)
}
\arguments{
\item{RegData}{NGER-data, dataramme}
}
\value{
Leverer formatert tabell for nøkkeltall
}
\description{
Nøkkeltallstabell for laparoskopi
}
