% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appNGER.R
\name{ui_nger}
\alias{ui_nger}
\title{Brukergrensesnitt (ui) til nger-appen}
\usage{
ui_nger()
}
\value{
Brukergrensesnittet (ui) til nger-appen
}
\description{
Brukergrensesnitt (ui) til nger-appen
}
