% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERtabeller.R
\name{visVanligsteProcDiag}
\alias{visVanligsteProcDiag}
\title{Vise vanligste prosedyrer eller diagnoser}
\usage{
visVanligsteProcDiag(RegData, ant = 20, prosdiag = "pros")
}
\arguments{
\item{RegData}{dataramme}

\item{ant}{antall prosedyrer/diagnoser}

\item{prosdiag}{'pros'-prosedyrer, 'diag'-diagnoser}
}
\description{
Vise vanligste prosedyrer eller diagnoser
}
