% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERhjelpefunksjoner.R
\name{dataTilOffVisning}
\alias{dataTilOffVisning}
\title{Generere data til SKDEs interaktive nettsider}
\usage{
dataTilOffVisning(
  RegData = RegData,
  valgtVar,
  aggData = 0,
  datoFra = "2014-01-01",
  datoTil = Sys.Date(),
  indID = "indDummy",
  lastNedFil = 0,
  filUt = "dummy"
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{- parameter for hvilken variabel det skal lages resultat for}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{indID}{indikator-id, eks. 'ind1', 'ind2', osv.}

\item{filUt}{tilnavn for utdatatabell (fjern?)}
}
\value{
Datafil til Resultatportalen. Utfila må inneholde:
'year', 'orgnr', 'var', 'denominator', 'caregiver' og 'ind_id'
}
\description{
Generere data til SKDEs interaktive nettsider
}
