% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERhjelpefunksjoner.R
\name{dataTilOffVisning}
\alias{dataTilOffVisning}
\title{Generere data til Resultatportalen/SKDE-viser}
\usage{
dataTilOffVisning(
  RegData = RegData,
  valgtVar,
  datoFra = "2014-01-01",
  aar = 0,
  OpMetode = 0,
  indID = "indDummy",
  ResPort = 0,
  lastNedFil = 0,
  filUt = "dummy"
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{- beinsmLavPre, peropKompDura, sympVarighUtstr, p.t. 10 kvalitetsind.}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{OpMetode}{1: Laparoskopi
2: Hysteroskopi
3: Begge
4: LCD01 eller LCD04 (total laparoskopisk hysterektomi)
5: LCC11 (laparoskopisk subtotal hysterektomi)
6: LCD11 (laparoskopisk assistert vaginal hysterektomi)
7: Robotassisert inngrep
8: Kolpopeksiene}

\item{indID}{indikator-id, eks. 'ind1', 'ind2', osv.}

\item{ResPort}{1-hvis data til resultatportalen (standard), 0-data til SKDE-viser}

\item{filUt}{tilnavn for utdatatabell (fjern?)}
}
\value{
Datafil til Resultatportalen
}
\description{
Generere data til Resultatportalen/SKDE-viser
}
