% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERFigAndeler.R
\name{FigAndeler}
\alias{FigAndeler}
\title{Lag søylediagram eller stabelplott som viser andeler av ulike variabler}
\usage{
FigAndeler(RegData, valgtVar, datoFra = "2013-01-01",
  datoTil = "2050-12-31", minald = 0, maxald = 130, outfile = "",
  reshID, enhetsUtvalg = 1, MCEType = 99, hentData = 0,
  preprosess = TRUE)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel skal visualiseres
Alder: Pasientens alder, 5-årige aldersgrupper
Education: Pasientens utdanning
Komplikasjoner
LapAccessMethod: Teknikk for laparaskopisk tilgang
LapNumHjelpeinnstikk: Antall hjelpeinnstikk
MaritalStatus: Sivilstand
MCEType: Operasjonsmetode
PatientNorwegian: Pasientens norskkunnskaper
OpAnesthetic: Anestesitype
OpBMICategory: BMI-kategori
Opcat: Hastegrad av operasjon
OpEarlierVaginal: Tidligere vaginale inngrep
OpEarlierLaparoscopy: Tidligere laparoskopi
OpEarlierLaparatomy: Tidligere laparatomi
OpOpcatOutsideDaytime: Operasjon i legens vakttid
OpDaySurgery: Dagkirurgiske inngrep
OpType: Primæroperasjon eller reoperasjon}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Standard)
2: Egen enhet}

\item{MCEType}{1: Laparoskopi
2: Hysteroskopi
3: Begge
99: Alle}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data - foreslår vi alltid gjør dette. Ellers må vi ta høyde for det i funksjonen også
FALSE: Nei
TRUE: Ja (Standard)}
}
\value{
En figur med søylediagram (eller et stabelplot) av ønsket variabel
}
\description{
Denne funksjonen lager et søylediagram ellersom viser andeler av valgt variabel
filtrert på de utvalg som er gjort.
}

