% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERFigAndeler.R
\name{FigAndeler}
\alias{FigAndeler}
\title{Søylediagram som viser andeler av ulike variabler}
\usage{
FigAndeler(RegData = 0, valgtVar, datoFra = "2013-01-01",
  datoTil = "2050-12-31", minald = 0, maxald = 130, outfile = "",
  reshID, enhetsUtvalg = 1, MCEType = 99, Hastegrad = "",
  AlvorlighetKompl = "", hentData = 0, preprosess = TRUE)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres
Alder: Pasientens alder, 5-årige aldersgrupper
Education: Pasientens utdanning (1:Grunnskole, 2:VG, 3:Fagskole, 4:Universitet<4 år, 5:Universitet>4 år, 6:Ukjent)
FollowupSeriousness: Alvorlighetsgrad, postoperative komplikasjoner
    Kode 1-Lite alvorlig, 2-Middels alvorlig, 3-Alvorlig, 4-Dødelig
HypCompleteness: Gjennomføringsgrad av hysteroskopi
       Koder:    1-Fullstendig, 2-Ufullstendig, 3-Mislykket
HypComplications: Hysteroskopi intrapoerative komplikasjoner
KomplPost: Postoperative komplikasjoner
KomplPostUtd: Postoperative komplikasjoner for ulike utdanningsgrupper
KomplReopUtd: Andel reoperasjoner som følge av komplikasjon for ulike utdanningsgrupper
LapAccessMethod: Teknikk for laparaskopisk tilgang
LapComplications: Laparoskopiske intrapoerative komplikasjoner
LapEkstrautstyr: Laparaskopisk ekstrautstyr - Kommer, NY variabel: koagulasjon og klipping
LapIntraAbdominal: Laparoskopiske intraabdominale komplikasjoner
LapNumHjelpeinnstikk: Antall hjelpeinnstikk
MaritalStatus: Sivilstand
MCEType: Operasjonsmetode
PatientNorwegian: Pasientens norskkunnskaper
OpAnesthetic: Anestesitype
OpASA: ASA-grad
OpBMICategory: BMI-kategori
Opcat: Hastegrad av operasjon
OpDaySurgery: Dagkirurgiske inngrep
OpEarlierVaginal: Tidligere vaginale inngrep
OpEarlierLaparoscopy: Tidligere laparoskopi
OpEarlierLaparatomy: Tidligere laparatomi
OpOpcatOutsideDaytime: Operasjon i legens vakttid
OpType: Primæroperasjon eller reoperasjon}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Standard)
2: Egen enhet}

\item{MCEType}{1: Laparoskopi
2: Hysteroskopi
3: Begge
99: Alle}

\item{Hastegrad}{Hastegrad av operasjon.
1: Elektiv
2: Akutt
3: Ø-hjelp}

\item{AlvorlighetKompl}{Alvorlighetsgrad for postoperative komplikasjoner (Flervalg)
                         Angis som en vektor av tall som tekst, f.eks. c('1','2')
                         1: Lite alvorlig
                         2: Middels alvorlig
                         3: Alvorlig
                         4: Dødelig

Detajer: Her bør man liste opp hvilke variable funksjonen benytter.}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
FALSE: Nei
TRUE: Ja (Standard)}
}
\value{
En figur med søylediagram (fordeling) av ønsket variabel
}
\description{
Denne funksjonen lager et søylediagram som viser andeler (fordeling) av valgt variabel
filtrert på de utvalg som er gjort.
}

