% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERRegDataSQL.R, R/NGERhentData.R
\name{hentDataTabell}
\alias{hentDataTabell}
\title{Hent datatabell fra ngers database}
\usage{
hentDataTabell(
  tabellnavn = "operation",
  qVar = "*",
  datoFra = "2011-01-01",
  datoTil = Sys.Date(),
  egneVarNavn = 1
)

hentDataTabell(
  tabellnavn = "operation",
  qVar = "*",
  datoFra = "2011-01-01",
  datoTil = Sys.Date(),
  egneVarNavn = 1
)
}
\arguments{
\item{tabellnavn}{Navn på tabell som skal lastes inn.}

\item{egneVarNavn}{0 - Qreg-navn benyttes.
1 - selvvalgte navn fra Friendlyvar benyttes}
}
\description{
Hent datatabell fra ngers database

Hent datatabell fra ngers database
}
