% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERhjelpefunksjoner.R
\name{henteSamlerapporter}
\alias{henteSamlerapporter}
\title{Generere samlerapporter i app}
\usage{
henteSamlerapporter(
  filnavn,
  rnwFil,
  reshID = 0,
  datoFra = Sys.Date() - 180,
  datoTil = Sys.Date()
)
}
\arguments{
\item{filnavn}{benyttes i downloadhandler}

\item{rnwFil}{Fila som skal kompileres. Eks. 'Eksempel.rnw'}

\item{reshID}{reshID}

\item{datoFra}{startdato}

\item{datoTil}{sluttdato}
}
\description{
Generere samlerapporter i app
}
