% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERhjelpefunksjoner.R
\name{lageTulleData}
\alias{lageTulleData}
\title{Lage tulledata (simulerte data)}
\usage{
lageTulleData(
  RegData = 0,
  varBort = NA,
  datoFra = "2017-01-01",
  antSh = 26,
  antObs = 20000
)
}
\arguments{
\item{RegData}{dataramme}

\item{varBort}{variable som skal fjernes. Variabelnavn angis som vektor av tekststrenger}

\item{antSh}{antall sykehus}

\item{antObs}{antall observasjoner (rader)}
}
\description{
Lage tulledata (simulerte data)
}
