% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERLibUtvalg.R
\name{LibUtvalg}
\alias{LibUtvalg}
\title{Funksjon som gjør utvalg av dataene, returnerer det filtrerte datasettet og utvalgsteksten.}
\usage{
LibUtvalg(RegData, datoFra, datoTil, fargepalett = "BlaaOff", minald, maxald,
  MCEType)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{fargepalett}{Hvilken fargepalett skal brukes i figurer (Default: BlaaRapp)}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{MCEType}{1: Laparoskopi
2: Hysteroskopi
3: Begge
99: Alle}
}
\value{
UtData En liste bestående av det filtrerte datasettet, utvalgstekst for figur og tekststreng som angir fargepalett
}
\description{
Funksjon som gjør utvalg av dataene, returnerer det filtrerte datasettet og utvalgsteksten.
}

