% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERFigAndeler.R
\name{NGERFigAndeler}
\alias{NGERFigAndeler}
\title{Søylediagram som viser andeler av ulike variabler}
\usage{
NGERFigAndeler(
  RegData = 0,
  valgtVar,
  datoFra = "2013-01-01",
  datoTil = "2050-12-31",
  minald = 0,
  maxald = 130,
  outfile = "",
  reshID = 0,
  enhetsUtvalg = 0,
  OpMetode = 99,
  Hastegrad = "",
  AlvorlighetKompl = "",
  velgAvd = 0,
  velgDiag = 0,
  hentData = 0,
  preprosess = 1,
  ...
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Standard)
2: Egen enhet}

\item{OpMetode}{1: Laparoskopi
2: Hysteroskopi
3: Begge
4: LCD01 eller LCD04 (total laparoskopisk hysterektomi)
5: LCC11 (laparoskopisk subtotal hysterektomi)
6: LCD11 (laparoskopisk assistert vaginal hysterektomi)
7: Robotassisert inngrep
8: Kolpopeksiene}

\item{Hastegrad}{Hastegrad av operasjon.
1: Elektiv
2: Akutt
3: Ø-hjelp}

\item{AlvorlighetKompl}{Alvorlighetsgrad for postoperative komplikasjoner (Flervalg)
Angis som en vektor av tall som tekst, f.eks. c('1','2')
1: Lite alvorlig
2: Middels alvorlig
3: Alvorlig
4: Dødelig}

\item{velgAvd}{Velge hvilke avdelinger som skal vises i figurer med avdelingsvise resultater.
IKKE tatt høyde for sammenlikning mot "resten".}

\item{velgDiag}{0: Alle
1: Ovarialcyster (N83.0, N83.1, N83.2 og D27)
2: Endometriose, livmorvegg (N80.0)
3: Endometriose, unntatt livmorvegg.}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
0: Nei (Standard)
1: Ja}
}
\value{
En figur med søylediagram (fordeling) av ønsket variabel
}
\description{
Denne funksjonen lager et søylediagram som viser andeler (fordeling) av valgt variabel
filtrert på de utvalg som er gjort.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item Alder: Pasientens alder, 5-årige aldersgrupper
    \item Diagnoser: Hyppigst forekommende diagnoser
    \item HysGjforingsGrad: Gjennomføringsgrad av hysteroskopi
   		Koder:	1-Fullstendig, 2-Ufullstendig, 3-Mislykket
    \item HysKomplikasjoner: Hysteroskopi intrapoerative komplikasjoner
    \item KomplPostop: Postoperative komplikasjoner
    \item KomplPostUtd: Postoperative komplikasjoner for ulike utdanningsgrupper
    \item KomplReopUtd: Andel reoperasjoner som følge av komplikasjon for ulike utdanningsgrupper
    \item LapKomplikasjoner: Laparoskopiske intrapoerative komplikasjoner
    \item LapEkstrautstyr: Laparaskopisk ekstrautstyr - Kommer, NY variabel: koagulasjon og klipping
    \item LapIntraabdominell: Laparoskopiske intraabdominale komplikasjoner
    \item LapNumHjelpeinnstikk: Antall hjelpeinnstikk
    \item LapTeknikk: Laparaskopisk tilgang, teknikk og metode (Tidl LapTilgangsMetode)
    \item Norsktalende: Pasientens norskkunnskaper
    \item OpAnestesi: Anestesitype
    \item OpASA: ASA-grad
    \item OpBMI: BMI-kategori
    \item Opf0AlvorlighetsGrad: Alvorlighetsgrad, postoperative komplikasjoner
		  Kode 1-Lite alvorlig, 2-Middels alvorlig, 3-Alvorlig, 4-Dødelig
	 \item Opf0KomplInfeksjon: Type infeksjoner
	 \item Opf0metode: Opfølgingsmetode
    \item OpDagkirurgi: Dagkirurgiske inngrep
    \item OpIVaktTid: Operasjon i legens vakttid
    \item OpKategori: Hastegrad av operasjon
    \item OpMetode: Operasjonsmetode
    \item OpTid: Operasjonstid (minutter)
    \item OpTidlVagInngrep: Tidligere vaginale inngrep
    \item OpTidlLapsko: Tidligere laparoskopi
    \item OpTidlLaparotomi: Tidligere laparatomi
    \item OpType: Primæroperasjon eller reoperasjon
    \item R0ScoreEmo:
    \item R0ScoreEnergy:
    \item R0ScoreGeneral:
    \item R0ScorePain:
    \item R0ScorePhys:
    \item R0ScoreRoleLmtPhy:
    \item R0ScoreRoleLmtEmo:
    \item R0ScoreSosial:
    \item RegForsinkelse: MANGLER BEREGNINGSVARIABEL Tid fra operasjon til ferdigstilt registrering
    \item Prosedyrer: Hyppigst forekommende prosedyrer
    \item SivilStatus: Sivilstand
    \item Tss2Mott: Hvordan du ble mottatt på avd.
    \item Tss2Behandling:
    \item Tss2Lytte:
    \item Tss2Behandlere:
    \item Tss2Enighet:
    \item Tss2Generelt:
    \item Utdanning: Pasientens utdanning (1:Grunnskole, 2:VG, 3:Fagskole, 4:Universitet<4 år, 5:Universitet>4 år, 6:Ukjent)
   }
}
