% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERFigGjsnGrVar.R
\name{NGERFigGjsnGrVar}
\alias{NGERFigGjsnGrVar}
\title{Søylediagram med gjennomsnitt/median for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
NGERFigGjsnGrVar(RegData, datoFra = "2013-01-01",
  datoTil = "3000-12-31", valgtVar, minald = 0, maxald = 130,
  OpMetode = 99, AlvorlighetKompl = 99, Hastegrad = 99,
  valgtMaal = "gjsn", hentData = 0, preprosess = 1,
  grVar = "ShNavn", velgAvd = "", velgDiag = 0, Ngrense = 10,
  medKI = 1, lagFig = 1, outfile = "")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{valgtVar}{Hvilken variabel som skal visualiseres}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{OpMetode}{1: Laparoskopi
2: Hysteroskopi
3: Begge
4: LCD01 eller LCD04 (total laparoskopisk hysterektomi)
5: LCC11 (laparoskopisk subtotal hysterektomi)
6: LCD11 (laparoskopisk assistert vaginal hysterektomi)}

\item{AlvorlighetKompl}{Alvorlighetsgrad for postoperative komplikasjoner (Flervalg)
Angis som en vektor av tall som tekst, f.eks. c('1','2')
1: Lite alvorlig
2: Middels alvorlig
3: Alvorlig
4: Dødelig}

\item{Hastegrad}{Hastegrad av operasjon.
1: Elektiv
2: Akutt
3: Ø-hjelp}

\item{valgtMaal}{'med' = median. Alt annet gir gjennomsnitt

Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item Alder: Liggetid
    \item RAND36 - alle dimensjoner
    \item Registreringsforsinkelse
    \item TSS2
   }}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
0: Nei (Standard)
1: Ja}

\item{velgAvd}{Velge hvilke avdelinger som skal vises i figurer med avdelingsvise resultater.
IKKE tatt høyde for sammenlikning mot "resten".}

\item{velgDiag}{0: Alle
1: Ovarialcyster (N83.0, N83.1, N83.2 og D27)
2: Endometriose, livmorvegg (N80.0)
3: Endometriose, unntatt livmorvegg.}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}
}
\value{
Søylediagram med gjennomsnitt/median av valgt variabel for hvert sykehus
}
\description{
Funksjon som genererer en figur med gjennomsnitt/median
for hvert sykehus og kan ta inn ulike numeriske variable.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus
}
