% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERFigGjsnGrVar.R
\name{NGERFigGjsnGrVar}
\alias{NGERFigGjsnGrVar}
\title{Søylediagram med gjennomsnitt/median for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
NGERFigGjsnGrVar(RegData, datoFra = "2013-01-01", datoTil = "3000-12-31",
  valgtVar, minald = 0, maxald = 130, MCEType = 99,
  AlvorlighetKompl = 99, Hastegrad = 99, valgtMaal = "Gjsn",
  hentData = 0, preprosess = 1, grVar = "ShNavn", Ngrense = 10,
  medKI = 1, lagFig = 1, outfile = "")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{valgtVar}{Hvilken variabel som skal visualiseres}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{MCEType}{1: Laparoskopi
2: Hysteroskopi
3: Begge
4: LCD01 eller LCD04 (total laparoskopisk hysterektomi)
5: LCC11 (laparoskopisk subtotal hysterektomi)
6: LCD11 (laparoskopisk assistert vaginal hysterektomi)
99: Alle}

\item{AlvorlighetKompl}{Alvorlighetsgrad for postoperative komplikasjoner (Flervalg)
                         Angis som en vektor av tall som tekst, f.eks. c('1','2')
                         1: Lite alvorlig
                         2: Middels alvorlig
                         3: Alvorlig
                         4: Dødelig

Detajer: Her bør man liste opp hvilke variable funksjonen benytter.}

\item{Hastegrad}{Hastegrad av operasjon.
1: Elektiv
2: Akutt
3: Ø-hjelp}

\item{valgtMaal}{'Med' = median. Alt annet gir gjennomsnitt

Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item Liggedogn: Liggetid
   }

Detajer: Her bør man liste opp hvilke variable funksjonen benytter.}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
0: Nei (Standard)
1: Ja}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}
}
\value{
Søylediagram med gjennomsnitt/median av valgt variabel for hvert sykehus
}
\description{
Funksjon som genererer en figur med gjennomsnitt/median
for hvert sykehus og kan ta inn ulike numeriske variable.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}
