% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERFigGjsnTid.R
\name{NGERFigGjsnTid}
\alias{NGERFigGjsnTid}
\title{Utvikling over tid for gjennomsnitt/median av valgt variabel}
\usage{
NGERFigGjsnTid(RegData, valgtVar = "alder", datoFra = "2011-01-01",
  datoTil = "3000-12-31", tidsenhet = "Aar", minald = 0,
  maxald = 110, reshID = 0, outfile = "", enhetsUtvalg = 0,
  valgtMaal = "Gjsn", preprosess = 1, hentData = 0)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Standard)
2: Egen enhet}

\item{valgtMaal}{'gjsn': gir middelverdi (standard)
'med': gir median}

\item{preprosess}{Preprosesser data
0: Nei (Standard)
1: Ja}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}
}
\value{
Linjediagram som viser utvikling over tid for valgt variabel
}
\description{
Figuren viser gjennomsnitt/median per år med konfidensintervall for valgt variabel.
I bakgrunn vises konfidensintervall for det man har valt å sammenlikne med.
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item alder: Pasientens alders
   }
}
