% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERFigKvalInd.R
\name{NGERFigKvalInd}
\alias{NGERFigKvalInd}
\title{Søylediagram som viser flere kvalitetsindikatorer}
\usage{
NGERFigKvalInd(
  RegData,
  reshID = 0,
  velgAvd = 0,
  datoFra = "2013-01-01",
  datoTil = Sys.Date(),
  valgtVar = "kvalInd",
  enhetsUtvalg = 0,
  minald = 0,
  maxald = 130,
  OpMetode = 99,
  Hastegrad = 99,
  dagkir = 9,
  hentData = 0,
  preprosess = 1,
  velgDiag = 0,
  Ngrense = 10,
  outfile = "",
  ...
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{velgAvd}{Velge hvilke avdelinger som skal vises i figurer med avdelingsvise resultater.
IKKE tatt høyde for sammenlikning mot "resten".}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{valgtVar}{Hvilken variabel som skal visualiseres}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Standard)
2: Egen enhet}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{OpMetode}{1: Laparoskopi
2: Hysteroskopi
3: Begge
4: LCD01 eller LCD04 (total laparoskopisk hysterektomi)
5: LCC11 (laparoskopisk subtotal hysterektomi)
6: LCD11 (laparoskopisk assistert vaginal hysterektomi)
7: Robotassisert inngrep
8: Kolpopeksiene}

\item{Hastegrad}{Hastegrad av operasjon.
1: Elektiv
2: Akutt
3: Ø-hjelp}

\item{dagkir}{Dagkirurgi? 0-nei, 1-ja}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}

\item{preprosess}{Preprosesser data
0: Nei (Standard)
1: Ja}

\item{velgDiag}{0: Alle
1: Ovarialcyster (N83.0, N83.1, N83.2 og D27)
2: Endometriose, livmorvegg (N80.0)
3: Endometriose, unntatt livmorvegg.}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}
}
\value{
Søylediagram samling av kvalitetsindikatorer
}
\description{
Funksjon som genererer en figurer med ei gruppe kvalitetsindikatorer
}
\details{
Argumentet \emph{valgtVar} har følgende valgmuligheter:
   \itemize{
    \item RAND0: Alle dimensjonene i RAND36 ved operasjonstidspunkt. Gjennomsnitt
    \item RAND1: Alle dimensjonene i RAND36 ett år etter operasjon. Gjennomsnitt
    \item TSS0: Alle sp?rsm?lene i TSS2 ved oppf?lging etter 4-6uker. Andel av beste svaralternativ
    \item kvalInd: Samling av kvalitetsindikatorer
   }
}
