% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERFigPrePost.R
\name{NGERFigPrePost}
\alias{NGERFigPrePost}
\title{Søylediagram som viser frekvens av valgt variabel, målt i to tidspunkter}
\usage{
NGERFigPrePost(
  RegData,
  valgtVar = "Antall",
  datoFra = "2012-04-01",
  datoTil = "2050-12-31",
  minald = 0,
  maxald = 110,
  erMann = "",
  outfile = "",
  reshID,
  enhetsUtvalg = 0,
  preprosess = 1,
  hentData = 0,
  ...
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel som skal visualiseres
Boligforh: Boligforhold ved innleggelse og 3 mnd etter slaget
Bosituasjon: Bosituasjon ved innleggelse og 3 mnd etter slaget
Toalett: Toalettbesok før innleggelse og 3 mnd etter slaget
Forflytning: Evne til forflytning før innleggelse og 3 mnd etter slaget}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Standardverdi: 0)}

\item{maxald}{Alder, til og med (Standardverdi: 130)}

\item{outfile}{Navn på fil figuren skrives til. Standard: '' (Figur skrives
til systemets standard utdataenhet (som regel skjerm))}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Standard)
2: Egen enhet}

\item{preprosess}{Preprosesser data
0: Nei (Standard)
1: Ja}

\item{hentData}{Gjør spørring mot database
0: Nei, RegData gis som input til funksjonen (Standard)
1: Ja}
}
\value{
Søylediagram som fordeling av valgt variabel, målt både før og etter hjerneslaget
}
\description{
Funksjon som genererer en figur med som viser endring i en variabels fordeling fra før til etter.
Man kan velge om det skal vises andeler eller antall
}
\details{
Detajer: Her bør man liste opp hvilke variable funksjonen benytter...
}
