% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERgmlAre.R
\name{NGERHentRegDataVI}
\alias{NGERHentRegDataVI}
\title{Gamle tabeller fra Are
Provide dataframe for tab vi NGER
Provides NGER data for tab vi from staging}
\usage{
NGERHentRegDataVI(reportYear = 2099)
}
\value{
RegData data frame
}
\description{
Gamle tabeller fra Are
Provide dataframe for tab vi NGER
Provides NGER data for tab vi from staging
}
