% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERPreprosesser.R
\name{NGERPreprosess}
\alias{NGERPreprosess}
\title{Preprosesser data fra NGER}
\usage{
NGERPreprosess(RegData = RegData)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}
}
\value{
RegData En data.frame med det preprosesserte datasettet
}
\description{
Denne funksjonen definerer variabler og fjerner ikke-ferdigstilte registreringer
}

