% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NGERPreprosesser.R
\name{NGERPreprosess}
\alias{NGERPreprosess}
\title{Preprosesser data fra NGER}
\usage{
NGERPreprosess(RegData = RegData, reshID = reshID)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}
}
\value{
Data En list med det filtrerte datasettet og sykehusnavnet som tilsvarer reshID
}
\description{
Denne funksjonen definerer opp sammensatte variabler og fjerner ikke-ferdigstilte registreringer
}

