% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERhentData.R
\name{NGERRegDataSQL}
\alias{NGERRegDataSQL}
\title{Henter NGER-data}
\usage{
NGERRegDataSQL(
  datoFra = "2013-01-01",
  datoTil = Sys.Date(),
  medPROM = 1,
  gml = 0,
  ...
)
}
\arguments{
\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{medPROM:}{koble på RAND og TSS2-variabler}

\item{alleData}{1- alle variabler med, 0 - utvalgte variabler med}
}
\value{
RegData data frame
}
\description{
Henter NGER-data
}
