% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERRegDataSQL.R
\name{NGERRegDataSQL}
\alias{NGERRegDataSQL}
\title{Henter global dataramme for NGER}
\usage{
NGERRegDataSQL(datoFra = "2014-01-01", datoTil = Sys.Date(), medPROM = 1, ...)
}
\arguments{
\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{medPROM:}{koble på RAND og TSS2-variabler}
}
\value{
RegData data frame
}
\description{
Henter NGER-data fra staging.
Klargjort (delvis) for å kunne  sende inn en vektor med hvilke variable det skal spørres etter.
}
