% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERtabeller.R
\name{NGERtabeller}
\alias{NGERtabeller}
\alias{tabAntOpphShMnd}
\alias{tabAntOpphSh5Aar}
\alias{tabAntOpphPasSh5Aar}
\alias{lagTabavFig}
\title{Funksjoner for å lage tabeller Group of functions page title}
\usage{
tabAntOpphShMnd(RegData, datoTil, antMnd = 6)

tabAntOpphSh5Aar(RegData, datoTil)

tabAntOpphPasSh5Aar(RegData, gr = "opph", datoTil)

lagTabavFig(UtDataFraFig)
}
\arguments{
\item{RegData}{data}

\item{datoTil}{sluttdato. Brukes i tabellene AntOpph per 12 mnd og Belegg}

\item{personIDvar}{Variabelen som angir pasientidentifikasjon}
}
\value{
Div tabeller
}
\description{
Fil som beregner div tabeller.Group of functions Description section
}
\details{
Detaljer. kommer senereGroup of functions Details paragraph.

Fil som inneholder funksjoner for å lage tabeller, i første rekke tellinger av personer
RegData må inneholde InnDato og Aar.
Aktuelle tabeller:
-tabAntOpphSh12mnd: Antall opphold per måned og enhet siste 12 måneder fram til datoTil.
-tabAntOpphSh5Aar:Antall opphold per år og enhet siste 5 år (inkl. inneværende år) fram til datoTil.
}
\section{Antall opphold siste X (antMnd) mnd}{
NA
}

\section{Antall opphold siste 5 år}{
NA
}

\section{Antall registreringer/pasienter siste 5 år}{

Hmmm
}

\section{Vise figurdata som tabell}{
NA
}

