% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERtabeller.R
\name{NGERtabeller}
\alias{NGERtabeller}
\alias{tabAntOpphShMnd}
\alias{tabAntOpphSh5Aar}
\alias{tabAntSkjema}
\alias{lagTabavFig}
\alias{lagTabavFigGjsnGrVar}
\alias{NGERpasientegenskaper}
\alias{instrumentbruk}
\alias{KomplLapTab}
\alias{konvertertLap}
\alias{visVanligsteProcDiag}
\title{Funksjoner for å lage tabeller Group of functions page title}
\usage{
tabAntOpphShMnd(RegData, datoTil = Sys.Date(), antMnd = 6,
  reshID = 0)

tabAntOpphSh5Aar(RegData, datoTil = Sys.Date())

tabAntSkjema(SkjemaOversikt, datoFra = "2019-01-01",
  datoTil = Sys.Date(), skjemastatus = 1)

lagTabavFig(UtDataFraFig, figurtype = "andeler")

lagTabavFigGjsnGrVar(UtDataFraFig)

NGERpasientegenskaper(RegData, datoFra = "2019-01-01",
  datoTil = Sys.Date(), tidsenhet = "Kvartal")

instrumentbruk(RegData, datoFra = "2019-01-01", datoTil = Sys.Date())

KomplLapTab(RegData, reshID = 0, datoFra = "2019-01-01",
  datoTil = Sys.Date())

konvertertLap(RegData, reshID = 0, datoFra = "2016-01-01",
  datoTil = Sys.Date())

visVanligsteProcDiag(RegData, ant = 20, prosdiag = "pros")
}
\arguments{
\item{RegData}{data}

\item{datoTil}{sluttdato. Brukes i tabellene AntOpph per 12 mnd og Belegg}

\item{personIDvar}{Variabelen som angir pasientidentifikasjon}
}
\value{
Div tabeller
}
\description{
Fil som beregner div tabeller.Group of functions Description section
}
\details{
Detaljer. kommer senereGroup of functions Details paragraph.

Fil som inneholder funksjoner for å lage tabeller, i første rekke tellinger av personer
RegData må inneholde InnDato og Aar.
Aktuelle tabeller:
-tabAntOpphSh12mnd: Antall opphold per måned og enhet siste 12 måneder fram til datoTil.
-tabAntOpphSh5Aar:Antall opphold per år og enhet siste 5 år (inkl. inneværende år) fram til datoTil.
}
\section{Antall opphold siste X (antMnd) mnd}{
NA
}

\section{Antall opphold siste 5 år}{
NA
}

\section{Hvor mange skjema av hver type}{
NA
}

\section{Vise figurdata som tabell}{
NA
}

\section{Vise figurdata som tabell, sentralmål per sykshus}{
NA
}

\section{Generere tabell med nøkkeltall}{
NA
}

\section{instrumentbruk, Laparaskopi}{
NA
}

\section{komplikasjoner, Laparaskopi}{
NA
}

\section{Konvertert laparoskopi til laparatomi}{
NA
}

\section{Vise vanligste prosedyrer eller diagnoser}{
NA
}

