% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERgmlAre.R
\name{NGERtabXXI}
\alias{NGERtabXXI}
\title{Generate tab XXI}
\usage{
NGERtabXXI(RegData)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}
}
\value{
list $tabXXI data frame of table data

list $personsWithMultipleCompl number of persons
}
\description{
Yes, generate tab XXI
}
