% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NGERtabXXI.R
\name{NGERtabXXI}
\alias{NGERtabXXI}
\title{Generate tab XXI}
\usage{
NGERtabXXI(datoFra = "2014-01-01", datoTil = "2014-04-01")
}
\arguments{
\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}
}
\value{
dataframe tabXXI
}
\description{
Yes, generate tab XXI, uhuu
}

