% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERFigAndelerGrVar.R
\name{PlotAndelerGrVar}
\alias{PlotAndelerGrVar}
\title{Title}
\usage{
PlotAndelerGrVar(
  RegData,
  Variabel,
  grVar = "ShNavn",
  hovedgrTxt = "",
  KvalIndGrenser = NA,
  tittel = "tittel",
  utvalgTxt = "",
  Ngrense = 10,
  titleSize = 20,
  subtitleSize = 15,
  legendSize = 12,
  axisTextSize = 10,
  bestKvalInd = "lav",
  nTicks = 5,
  fargepalett = "BlaaOff",
  outfile = ""
)
}
\arguments{
\item{RegData}{dataramme}

\item{Variabel}{var}

\item{grVar}{grvar}

\item{hovedgrTxt}{hovedtxt}

\item{KvalIndGrenser}{KIgrenser}

\item{fargepalett}{Hvilken fargepalett skal brukes i figurer (Default: BlaaRapp)}

\item{outfile}{}
}
\description{
Title
}
