% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERtabeller.R
\name{rader}
\alias{rader}
\title{Definerer rader i en nøkkeltallstabell}
\usage{
rader(RegData, var, stat = "median", verdi = 1, met = "hys")
}
\arguments{
\item{RegData}{- dataramme, NGER-data}

\item{var}{- hvilken variabel en ønsker resultat for}

\item{stat}{- 'median' eller 'pst' (prosent)}

\item{verdi}{- hvilken verdi/kode som skal telles når en beregner prosent}

\item{met}{- 'hys' (hysteroskopi) eller 'lap' (laparoskopi)}
}
\value{
Leverer beregnede tall til rad i nøkkeltallstabell
}
\description{
Definerer rader i en nøkkeltallstabell
}
