% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGERtabeller.R
\name{tabAntSkjema}
\alias{tabAntSkjema}
\title{Hvor mange skjema av hver type}
\usage{
tabAntSkjema(
  SkjemaOversikt,
  datoFra = "2019-01-01",
  datoTil = Sys.Date(),
  skjemastatus = 1
)
}
\description{
Hvor mange skjema av hver type
}
